Refs 9393. Convert template to class hierarchy
