+++ src/cpp/rtps/history/TopicPayloadPool_impl/DynamicReusable.hpp
@@ -16,79 +16,49 @@
  * @file DynamicReusable.hpp
  */
 
+#ifndef RTPS_HISTORY_TOPICPAYLOADPOOLIMPL_DYNAMIC_REUSABLE_HPP
+#define RTPS_HISTORY_TOPICPAYLOADPOOLIMPL_DYNAMIC_REUSABLE_HPP
+
+#include <rtps/history/TopicPayloadPool.hpp>
+
 namespace eprosima {
 namespace fastrtps {
 namespace rtps {
-namespace TopicPayloadPool {
 
-template <>
-class Impl<DYNAMIC_REUSABLE_MEMORY_MODE> : public BaseImpl
+class DynamicReusableTopicPayloadPool : public TopicPayloadPool
 {
     bool get_payload(
             uint32_t size,
             CacheChange_t& cache_change) override
     {
-        octet* payload = nullptr;
-        if (free_payloads_.empty())
+        if (!TopicPayloadPool::get_payload(size, cache_change))
         {
-            payload = allocate(size); //Allocates a single payload
-            cache_change.serializedPayload.data = payload;
-            if (payload != nullptr)
-            {
-                cache_change.serializedPayload.max_size = size;
-                cache_change.payload_owner(this);
-                return true;
-            }
-            cache_change.serializedPayload.max_size = 0;
-            cache_change.payload_owner(nullptr);
             return false;
         }
 
-        octet* data = free_payloads_.back().data;
-        uint32_t max_size = free_payloads_.back().max_size;
-
         // Resize if needed
-        if (size > max_size)
+        if (size > cache_change.serializedPayload.max_size)
         {
-            octet* old_data = data;
-            data = (octet*)realloc(data, size);
-            if (!data)
+            if (!resize_payload(cache_change.serializedPayload.data, cache_change.serializedPayload.max_size, size))
             {
-                // Nothing changed on the buffers, so nothing to undo
-                throw std::bad_alloc();
-            }
-            memset(data + max_size, 0, (size - max_size) * sizeof(octet));
-            max_size = size;
-
-            // Find data in allPayloads to update the pointer
-            std::vector<octet*>::iterator target =
-                    std::find(all_payloads_.begin(), all_payloads_.end(), old_data);
-            if (target != all_payloads_.end())
-            {
-                *target = data;
-            }
-            else
-            {
-                logError(RTPS_HISTORY, "Found a free payload that is not logged in the Pool");
+                logError(RTPS_HISTORY, "Failed to resize the payload");
+                release_payload(cache_change);
                 return false;
             }
         }
-
-        free_payloads_.pop_back();
-        cache_change.serializedPayload.data = data;
-        cache_change.serializedPayload.max_size = max_size;
-        cache_change.payload_owner(this);
 
         return true;
     }
 
     bool release_history(
             const PoolConfig& config,
-            bool /*is_reader*/) override
+            bool is_reader) override
     {
-        assert(config.memory_policy == memory_policy());
-
-        update_maximum_size(config, false);
+        if (!TopicPayloadPool::release_history(config, is_reader))
+        {
+            return false;
+        }
+        
         return shrink(max_pool_size_);
     }
 
@@ -101,7 +71,8 @@ protected:
 
 };
 
-}  // namespace TopicPayloadPool
 }  // namespace rtps
 }  // namespace fastrtps
 }  // namespace eprosima
+
+#endif  // RTPS_HISTORY_TOPICPAYLOADPOOLIMPL_DYNAMIC_REUSABLE_HPP
