+++ src/cpp/rtps/history/TopicPayloadPool_impl/PreallocatedWithRealloc.hpp
@@ -16,17 +16,20 @@
  * @file Preallocated.hpp
  */
 
+#ifndef RTPS_HISTORY_TOPICPAYLOADPOOLIMPL_PREALLOCATED_REALLOC_HPP
+#define RTPS_HISTORY_TOPICPAYLOADPOOLIMPL_PREALLOCATED_REALLOC_HPP
+
+#include <rtps/history/TopicPayloadPool.hpp>
+
 namespace eprosima {
 namespace fastrtps {
 namespace rtps {
-namespace TopicPayloadPool {
 
-template <>
-class Impl<PREALLOCATED_WITH_REALLOC_MEMORY_MODE> : public BaseImpl
+class PreallocatedReallocTopicPayloadPool : public TopicPayloadPool
 {
 public:
 
-    explicit Impl(
+    explicit PreallocatedReallocTopicPayloadPool(
             uint32_t payload_size)
         : min_payload_size_(payload_size)
         , minimum_pool_size_for_writers(0)
@@ -38,67 +41,34 @@ public:
             uint32_t size,
             CacheChange_t& cache_change) override
     {
-        octet* payload = nullptr;
-        if (free_payloads_.empty())
+        if (!TopicPayloadPool::get_payload(std::max(size, min_payload_size_), cache_change))
         {
-            payload = allocate(std::max(size, min_payload_size_)); //Allocates a single payload
-            cache_change.serializedPayload.data = payload;
-            if (payload != nullptr)
-            {
-                cache_change.serializedPayload.max_size = std::max(size, min_payload_size_);
-                cache_change.payload_owner(this);
-                return true;
-            }
-            cache_change.serializedPayload.max_size = 0;
-            cache_change.payload_owner(nullptr);
             return false;
         }
 
-        octet* data = free_payloads_.back().data;
-        uint32_t max_size = free_payloads_.back().max_size;
-
         // Resize if needed
-        if (size > max_size)
+        if (size > cache_change.serializedPayload.max_size)
         {
-            octet* old_data = data;
-            data = (octet*)realloc(data, size);
-            if (!data)
+            if (!resize_payload(cache_change.serializedPayload.data, cache_change.serializedPayload.max_size, size))
             {
-                // Nothing changed on the buffers, so nothing to undo
-                throw std::bad_alloc();
-            }
-            memset(data + max_size, 0, (size - max_size) * sizeof(octet));
-            max_size = size;
-
-            // Find data in allPayloads to update the pointer
-            std::vector<octet*>::iterator target =
-                    std::find(all_payloads_.begin(), all_payloads_.end(), old_data);
-            if (target != all_payloads_.end())
-            {
-                *target = data;
-            }
-            else
-            {
-                logError(RTPS_HISTORY, "Found a free payload that is not logged in the Pool");
+                logError(RTPS_HISTORY, "Failed to resize the payload");
+                release_payload(cache_change);
                 return false;
             }
         }
 
-        free_payloads_.pop_back();
-        cache_change.serializedPayload.data = data;
-        cache_change.serializedPayload.max_size = max_size;
-        cache_change.payload_owner(this);
-
         return true;
     }
-    
+
     bool reserve_history(
             const PoolConfig& config,
             bool is_reader) override
     {
-        assert(config.memory_policy == memory_policy());
+        if (!TopicPayloadPool::reserve_history(config, is_reader))
+        {
+            return false;
+        }
 
-        update_maximum_size(config, true);
         if (is_reader)
         {
             minimum_pool_size_for_readers =
@@ -109,16 +79,19 @@ public:
             minimum_pool_size_for_writers += config.initial_size;
         }
 
-        return reserve(minimum_pool_size_for_writers + minimum_pool_size_for_readers, min_payload_size_);
+        reserve(minimum_pool_size_for_writers + minimum_pool_size_for_readers, min_payload_size_);
+        return true;
     }
 
     bool release_history(
             const PoolConfig& config,
-            bool /*is_reader*/) override
+            bool is_reader) override
     {
-        assert(config.memory_policy == memory_policy());
-
-        update_maximum_size(config, false);
+        if (!TopicPayloadPool::release_history(config, is_reader))
+        {
+            return false;
+        }
+        
         return shrink(max_pool_size_);
     }
 
@@ -136,7 +109,8 @@ private:
     uint32_t minimum_pool_size_for_readers;    //< Initial pool size due to readers (max of all readers)
 };
 
-}  // namespace TopicPayloadPool
 }  // namespace rtps
 }  // namespace fastrtps
 }  // namespace eprosima
+
+#endif  // RTPS_HISTORY_TOPICPAYLOADPOOLIMPL_PREALLOCATED_REALLOC_HPP
